Update alka.html
Made minor changes. It would be great if we could also look at a few examples of patents filed in the information area. Also, if the organization does not have an IPMS, what are the other options to privately file for a patent? Are there private entities like we have investors who will fund you to file a patent? You could also provide the example of http://witworks.com/. They are an interesting Indian startup who help bring your idea to life. Of course, read the fine print : )